+++ packages/subtree-updater/src/sync/subgraph/adapter.ts
@@ -12,6 +12,7 @@ import {
   TotalEntityIndexTrait,
   max,
   WithTotalEntityIndex,
+  range,
 } from "@nocturne-xyz/sdk";
 import { Insertion, SubtreeUpdaterSyncAdapter } from "../syncAdapter";
 import { fetchTreeInsertions, fetchLatestSubtreeIndex } from "./fetch";
@@ -33,7 +34,7 @@ export class SubgraphSubtreeUpdaterSyncAdapter
   iterInsertions(
     startTotalEntityIndex: TotalEntityIndex,
     opts?: IterSyncOpts
-  ): ClosableAsyncIterator<WithTotalEntityIndex<Insertion>> {
+  ): ClosableAsyncIterator<WithTotalEntityIndex<Insertion>[]> {
     const endpoint = this.graphqlEndpoint;
     const logger = this.logger;
     const endTotalEntityIndex = opts?.endTotalEntityIndex;
@@ -84,16 +85,17 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                   insertionKind: "zeros",
                   insertion: meta,
                 });
-              for (let i = 0; i < insertion.numZeros; i++) {
-                yield {
-                  inner: {
-                    noteCommitment: TreeConstants.ZERO_VALUE,
-                    merkleIndex: startIndex + i,
-                  },
-                  // HACK: add `i` to `totalEntityIndex` to ensure all of the zeros have unique `totalEntityIndex`s
-                  totalEntityIndex: totalEntityIndex + BigInt(i),
-                };
-              }
+
+              const batch = range(0, insertion.numZeros).map((i) => ({
+                inner: {
+                  noteCommitment: TreeConstants.ZERO_VALUE,
+                  merkleIndex: startIndex + i,
+                },
+                // HACK: add `i` to `totalEntityIndex` to ensure all of the zeros have unique `totalEntityIndex`s
+                totalEntityIndex: totalEntityIndex + BigInt(i),
+              }));
+
+              yield batch;
             } else if (NoteTrait.isEncryptedNote(insertion)) {
               const noteCommitment = (insertion as IncludedEncryptedNote)
                 .commitment;
@@ -108,13 +110,15 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                   meta,
                 });
 
-              yield {
+              const res = {
                 inner: {
                   noteCommitment,
                   merkleIndex: insertion.merkleIndex,
                 },
                 totalEntityIndex,
               };
+
+              yield [res];
             } else {
               const meta = {
                 merkleIndex: insertion.merkleIndex,
@@ -126,10 +130,12 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                   meta,
                 });
 
-              yield {
+              const res = {
                 inner: insertion as IncludedNote,
                 totalEntityIndex,
               };
+
+              yield [res];
             }
           }
         } else {